feat(start): allow user to define branch name templates (#47)
* fix(config): do not use empty object when updating local config file
* Now the issue types are fetched from Jira and stored in the config
file with a short name.  There are some defaults to match
the existing behavior (feature and story => f, task => c). In
any other cases, the first letter of the type is used for the short
name.
* Users can define the templates in the config file under jira =>
templates => branch. For now this is the data passed to those templates:
issue.shortName, issue.key and issue.sanitizedSummary. In order to
access them from the template, { and } must be used to interpolate.

Fixes #43, #38.