feat(Jira): add DONE status
Modify the infer process to not mix in progress and done.
Modify the manual process to ask for a done id.

BREAKING CHANGE:

Old clients don't have a DONE id in the config. There should be a
process that checks if this id exist so we can ask for it if it doesn't.