feat(start): add option to create an issue
If the -c option is specified, then it will create a new issue. Start
with `-c` needs to know the project `-p`, the type of issue to create
`-t` and a description of the issue.

Some extra validation rules needed to be added to the start script.
The jira client now has a method to create an issue.